+++ apps/evm/src/libs/errors/handleContractError/decodeWithContractErrorAbis/constants.ts
@@ -1,4 +1,4 @@
-import { type Abi, type Hex, decodeErrorResult } from 'viem';
+import type { Abi } from 'viem';
 
 import {
   isolatedPoolComptrollerAbi,
@@ -23,12 +23,10 @@ import {
   xvsTokenOmnichainAbi,
 } from 'libs/contracts';
 
-import type { ParsedContractError } from './parseContractError';
-
 // ABIs scanned to decode raw revert data when viem has not pre-decoded it.
 // Includes all Venus contracts the frontend interacts with, plus third-party
 // contracts users can reach (smart accounts, swap, bridge).
-const CONTRACT_ERROR_ABIS: Abi[] = [
+export const CONTRACT_ERROR_ABIS: Abi[] = [
   // Venus — core lending
   isolatedPoolComptrollerAbi,
   legacyPoolComptrollerAbi,
@@ -54,18 +52,3 @@ const CONTRACT_ERROR_ABIS: Abi[] = [
   nexusAccountFactoryAbi,
   nexusBoostrapAbi,
 ];
-
-export const decodeWithContractErrorAbis = (
-  rawData: Hex,
-  signature: Hex,
-): ParsedContractError | undefined => {
-  for (const abi of CONTRACT_ERROR_ABIS) {
-    try {
-      const decoded = decodeErrorResult({ abi, data: rawData });
-      return { errorName: decoded.errorName, args: decoded.args, signature };
-    } catch {
-      // selector not in this ABI
-    }
-  }
-  return undefined;
-};
