+++ apps/evm/src/libs/errors/handleContractError/readRawRevertData/index.ts
@@ -1,6 +1,6 @@
 import type { BaseError, Hex } from 'viem';
 
-import { isHexSelector } from './isHexSelector';
+import { isHexSelector } from '../isHexSelector';
 
 export const readRawRevertData = (error: BaseError): Hex | undefined => {
   const layer = error.walk(e => isHexSelector((e as { data?: unknown } | null)?.data));
